+++ test/unit/typedQueryBuilderTests.ts
@@ -1030,12 +1030,12 @@ describe('TypedKnexQueryBuilder', () => {
 
     it('should create query with where null', done => {
         const typedKnex = new TypedKnex(knex({ client: 'postgresql' }));
-        const query = typedKnex.query(User).whereNull(c => c.name);
+        const query = typedKnex.query(User).whereNull(c => c.name).orWhereNull(c => c.name);
 
         const queryString = query.toQuery();
         assert.equal(
             queryString,
-            'select * from "users" where "users"."name" is null'
+            'select * from "users" where "users"."name" is null or "users"."name" is null'
         );
 
         done();
@@ -1043,12 +1043,12 @@ describe('TypedKnexQueryBuilder', () => {
 
     it('should create query with where not null', done => {
         const typedKnex = new TypedKnex(knex({ client: 'postgresql' }));
-        const query = typedKnex.query(User).whereNotNull(c => c.name);
+        const query = typedKnex.query(User).whereNotNull(c => c.name).orWhereNotNull(c => c.name);
 
         const queryString = query.toQuery();
         assert.equal(
             queryString,
-            'select * from "users" where "users"."name" is not null'
+            'select * from "users" where "users"."name" is not null or "users"."name" is not null'
         );
 
         done();